Agrego cambios en pantalla crud espacios
